+++ src/main.ts
@@ -1,74 +1,4 @@
-const isArraySorted = require('is-array-sorted')
-
-class AsyncQueue<T> {
-    queue: Array<T>
-
-    constructor() {
-        this.queue = Array<T>()
-    }
-
-    enqueue(element: T): void {
-        this.queue.push(element)
-    }
-
-    async dequeue(): Promise<T> {
-        return this.queue.shift()
-    }
-}
-
-class AsyncSemaphore {
-    private promises = Array<() => void>()
-
-    constructor(private permits: number) {}
-
-    signal(): void {
-        this.permits += 1
-        if (this.promises.length > 0) this.promises.pop()!()
-    }
-
-    async wait(): Promise<void> {
-        this.permits -= 1
-        if (this.permits < 0 || this.promises.length > 0)
-            await new Promise(r => this.promises.unshift(r))
-    }
-}
-
-async function testAsyncQueueBehavior(nOps: number): Promise<Boolean> {
-    const result = new Array<number>()
-    const q = new AsyncQueue<number>()
-
-    const enqueue = (m: number) => q.enqueue(m)
-    const dequeue = () => q.dequeue()
-    const promises = Array<Promise<void>>()
-
-    let enqueues = 0
-    let dequeues = 0
-
-    // Do a random permutation of enqueing and dequeing
-    for (let i = 0; i < nOps; i += 1) {
-        if (Math.random() > 0.5) {
-            enqueues += 1
-            // console.log(`${Date.now()} Enqueuing ${enqueues}`)
-            enqueue(enqueues)
-        } else {
-            dequeues += 1
-            // console.log(`${Date.now()} Dequeuing`)
-            promises.push(dequeue().then(v => { result.push(v) }))
-        }
-    }
-
-    // console.log(`Total enqueues ${enqueues}; dequeues ${dequeues}`)
-    const pending = Math.min(enqueues, dequeues)
-    await Promise.all(promises.slice(0, pending))
-
-    // Length should be equal minimum between enqueues and dequeues
-    const isLengthOk = pending === result.length 
-
-    // Messages should be ordered
-    const isSorted = isArraySorted(result)
-
-    return isLengthOk && isSorted
-}
+import { testAsyncQueueBehavior } from './exercises/exercise1'
 
 setInterval(() => { }, 1000); // run program forever until an explicit exit occurs
 (async () => {
